requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
docx2txt==0.9
executing==2.2.1
fire==0.7.1
Flask==3.1.2
fonttools==4.61.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.63.0
googleapis-common-protos==1.72.0
groq==1.0.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.2
opencv-python-headless==4.13.0.90
packaging==26.0
parso==0.8.5
pdf2docx==0.5.8
platformdirs==4.5.1
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.26.7
pyparsing==3.3.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pyzmq==27.1.0
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==9.1.4
termcolor==3.3.0
threadpoolctl==3.6.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.5.3
websockets==15.0.1
Werkzeug==3.1.5
